+++ test/client/karma.conf.js
@@ -3,20 +3,26 @@ module.exports = function(config) {
     // base path, that will be used to resolve files and exclude
     basePath: '../..',
 
-    frameworks: ['jasmine'],
+    frameworks: ['jasmine', 'commonjs'],
 
     // list of files / patterns to load in the browser
     files: [
-      'test/client/mocks.js',
-      'static/karma.src.js',
-      'test/client/*.spec.js'
+      'client/*.js',
+      'test/client/*.js'
     ],
 
     // list of files to exclude
-    exclude: [],
+    exclude: [
+      'client/main.js'
+    ],
+
+    preprocessors: {
+      'client/*.js': ['commonjs'],
+      'test/client/*.js': ['commonjs']
+    },
 
     // use dots reporter, as travis terminal does not support escaping sequences
-    // possible values: 'dots', 'progress', 'junit', 'teamcity'
+    // possible values: 'dots', 'progress'
     // CLI --reporters progress
     reporters: ['progress', 'junit'],
 
@@ -29,16 +35,12 @@ module.exports = function(config) {
     // CLI --port 9876
     port: 9876,
 
-    // cli runner port
-    // CLI --runner-port 9100
-    runnerPort: 9100,
-
     // enable / disable colors in the output (reporters and logs)
     // CLI --colors --no-colors
     colors: true,
 
     // level of logging
-    // possible values: karma.LOG_DISABLE || karma.LOG_ERROR || karma.LOG_WARN || karma.LOG_INFO || karma.LOG_DEBUG
+    // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
     // CLI --log-level debug
     logLevel: config.LOG_INFO,
 
@@ -55,11 +57,11 @@ module.exports = function(config) {
     // - PhantomJS
     // - IE (only Windows)
     // CLI --browsers Chrome,Firefox,Safari
-    browsers: [],
+    browsers: [process.env.TRAVIS ? 'Firefox' : 'Chrome'],
 
     // If browser does not capture in given timeout [ms], kill it
     // CLI --capture-timeout 5000
-    captureTimeout: 5000,
+    captureTimeout: 20000,
 
     // Auto run tests on start (when browsers are captured) and exit
     // CLI --single-run --no-single-run
@@ -69,16 +71,12 @@ module.exports = function(config) {
     // CLI --report-slower-than 500
     reportSlowerThan: 500,
 
-    // compile coffee scripts
-    preprocessors: {
-      '**/*.coffee': 'coffee'
-    },
-
     plugins: [
       'karma-jasmine',
       'karma-chrome-launcher',
       'karma-firefox-launcher',
-      'karma-junit-reporter'
+      'karma-junit-reporter',
+      'karma-commonjs'
     ]
   });
 };