+++ test/e2e/coverageQunit/karma.conf.js
@@ -9,7 +9,7 @@ module.exports = function(config) {
 
     autoWatch: true,
 
-    browsers: ['Chrome'],
+    browsers: [process.env.TRAVIS ? 'Firefox' : 'Chrome'],
 
     reporters: ['progress', 'coverage'],