+++ test/e2e/dojo/karma.conf.ignore.js
@@ -1,0 +1,64 @@
+module.exports = function(config) {
+  config.set({
+    // base path, that will be used to resolve files and exclude
+    basePath: '',
+
+    frameworks: ['jasmine', 'dojo'],
+
+    // list of files / patterns to load in the browser
+    files: [
+      'main.js',
+
+      // all the sources, tests
+      {pattern: '*.js', included: false}
+    ],
+
+
+    // test results reporter to use
+    // possible values: dots || progress
+    reporters: ['dots'],
+
+
+    // web server port
+    port: 9876,
+
+
+    // cli runner port
+    runnerPort: 9100,
+
+
+    // enable / disable colors in the output (reporters and logs)
+    colors: true,
+
+
+    // level of logging
+    // possible values: LOG_DISABLE || LOG_ERROR || LOG_WARN || LOG_INFO || LOG_DEBUG
+    logLevel: config.LOG_INFO,
+
+
+    // enable / disable watching file and executing tests whenever any file changes
+    autoWatch: true,
+
+
+    // Start these browsers, currently available:
+    // - Chrome
+    // - ChromeCanary
+    // - Firefox
+    // - Opera
+    // - Safari
+    // - PhantomJS
+    browsers: [process.env.TRAVIS ? 'Firefox' : 'Chrome'],
+
+
+    // Continuous Integration mode
+    // if true, it capture browsers, run tests and exit
+    singleRun: false,
+
+    plugins: [
+      'karma-dojo',
+      'karma-jasmine',
+      'karma-chrome-launcher',
+      'karma-firefox-launcher'
+    ]
+  });
+};