+++ test/e2e/pass-opts/karma.conf.js
@@ -1,0 +1,19 @@
+module.exports = function(config) {
+  config.set({
+    frameworks: ['jasmine'],
+
+    files: [
+      '*.js'
+    ],
+
+    browsers: [ process.env.TRAVIS ? 'Firefox' : 'Chrome' ],
+
+    reporters: ['dots'],
+
+    plugins: [
+      'karma-jasmine',
+      'karma-chrome-launcher',
+      'karma-firefox-launcher'
+    ],
+  });
+};