+++ test/e2e/pass-opts/test.js
@@ -1,0 +1,6 @@
+describe('config', function() {
+  it('should be passed through to the browser', function() {
+    expect(window.__karma__.config).toBeDefined();
+    expect(window.__karma__.config.args).toEqual(['arg1','arg2','arg3']);
+  });
+});