+++ test/e2e/sauce/karma.conf.js
@@ -1,0 +1,33 @@
+var TRAVIS_WITHOUT_SAUCE = process.env.TRAVIS_SECURE_ENV_VARS === 'false';
+
+module.exports = function(config) {
+  config.set({
+    frameworks: ['jasmine'],
+
+    files: [
+      '*.js'
+    ],
+
+    autoWatch: true,
+
+    browsers: [TRAVIS_WITHOUT_SAUCE ? 'Firefox' : 'sl_chrome_linux'],
+
+    reporters: ['dots'],
+
+    logLevel: config.LOG_DEBUG,
+
+    plugins: [
+      'karma-jasmine',
+      'karma-sauce-launcher',
+      'karma-firefox-launcher'
+    ],
+
+    customLaunchers: {
+      sl_chrome_linux: {
+        base: 'SauceLabs',
+        browserName: 'chrome',
+        platform: 'linux'
+      }
+    }
+  });
+};